zchore: move territory
