+++ api/src/pdc/middlewares/ScopeToGroupMiddleware.integration.spec.ts
@@ -1,15 +1,8 @@
-import {
-  afterAll,
-  assertEquals,
-  assertRejects,
-  beforeAll,
-  describe,
-  it,
-} from "@/dev_deps.ts";
+import { afterAll, assertEquals, assertRejects, beforeAll, describe, it } from "@/dev_deps.ts";
 import { ContextType, ForbiddenException } from "@/ilos/common/index.ts";
 import { PostgresConnection } from "@/ilos/connection-postgres/index.ts";
 import { env_or_default } from "@/lib/env/index.ts";
-import { TerritorySelectorsInterface } from "@/shared/territory/common/interfaces/TerritoryCodeInterface.ts";
+import { TerritorySelectorsInterface } from "@/pdc/services/territory/contracts/common/interfaces/TerritoryCodeInterface.ts";
 import { ScopeToGroupMiddleware } from "./ScopeToGroupMiddleware.ts";
 
 describe("ScopeToGroupMiddleware", () => {
@@ -101,8 +94,7 @@ describe("ScopeToGroupMiddleware", () => {
         territory_id: 1,
         authorizedZoneCodes: geo_selector,
       }),
-      (params: { geo_selector: TerritorySelectorsInterface }) =>
-        params.geo_selector,
+      (params: { geo_selector: TerritorySelectorsInterface }) => params.geo_selector,
       middlewareConfig,
     );
     assertEquals(result, { com: geo_selector.com });
@@ -116,8 +108,7 @@ describe("ScopeToGroupMiddleware", () => {
         territory_id: 1,
         authorizedZoneCodes: geo_selector,
       }),
-      (params: { geo_selector: TerritorySelectorsInterface }) =>
-        params.geo_selector,
+      (params: { geo_selector: TerritorySelectorsInterface }) => params.geo_selector,
       middlewareConfig,
     );
     assertEquals(result, { com: geo_selector.com });
