+++ api/src/pdc/services/territory/TerritoryServiceProvider.ts
@@ -2,15 +2,15 @@ import { ExtensionInterface, NewableType, serviceProvider } from "@/ilos/common/
 import { ServiceProvider as AbstractServiceProvider } from "@/ilos/core/index.ts";
 import { defaultMiddlewareBindings } from "@/pdc/providers/middleware/index.ts";
 import { ValidatorExtension, ValidatorMiddleware } from "@/pdc/providers/validator/index.ts";
-import { create } from "@/shared/territory/create.schema.ts";
-import { deleteTerritory } from "@/shared/territory/delete.schema.ts";
-import { binding as findBinding } from "@/shared/territory/find.schema.ts";
-import { binding as findGeoBySirenBinding } from "@/shared/territory/findGeoBySiren.schema.ts";
-import { binding as getAuthorizedCodesBinding } from "@/shared/territory/getAuthorizedCodes.schema.ts";
-import { binding as listBinding } from "@/shared/territory/list.schema.ts";
-import { binding as listGeoBinding } from "@/shared/territory/listGeo.schema.ts";
-import { patchContacts } from "@/shared/territory/patchContacts.schema.ts";
-import { update } from "@/shared/territory/update.schema.ts";
+import { create } from "@/pdc/services/territory/contracts/create.schema.ts";
+import { deleteTerritory } from "@/pdc/services/territory/contracts/delete.schema.ts";
+import { binding as findBinding } from "@/pdc/services/territory/contracts/find.schema.ts";
+import { binding as findGeoBySirenBinding } from "@/pdc/services/territory/contracts/findGeoBySiren.schema.ts";
+import { binding as getAuthorizedCodesBinding } from "@/pdc/services/territory/contracts/getAuthorizedCodes.schema.ts";
+import { binding as listBinding } from "@/pdc/services/territory/contracts/list.schema.ts";
+import { binding as listGeoBinding } from "@/pdc/services/territory/contracts/listGeo.schema.ts";
+import { patchContacts } from "@/pdc/services/territory/contracts/patchContacts.schema.ts";
+import { update } from "@/pdc/services/territory/contracts/update.schema.ts";
 import { FindGeoBySirenAction } from "./actions/geo/FindGeoBySirenAction.ts";
 import { IndexAllGeoAction } from "./actions/geo/IndexAllGeoAction.ts";
 import { ListGeoAction } from "./actions/geo/ListGeoAction.ts";
