+++ api/src/pdc/services/territory/actions/geo/FindGeoBySirenAction.ts
@@ -1,18 +1,15 @@
-import {
-  contentBlacklistMiddleware,
-  hasPermissionMiddleware,
-} from "@/pdc/providers/middleware/index.ts";
-import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { handler } from "@/ilos/common/index.ts";
+import { Action as AbstractAction } from "@/ilos/core/index.ts";
+import { contentBlacklistMiddleware, hasPermissionMiddleware } from "@/pdc/providers/middleware/index.ts";
 
-import { GeoRepositoryProviderInterfaceResolver } from "../../interfaces/GeoRepositoryProviderInterface.ts";
 import {
   handlerConfig,
   ParamsInterface,
   ResultInterface,
-} from "@/shared/territory/findGeoBySiren.contract.ts";
-import { alias } from "@/shared/territory/findGeoBySiren.schema.ts";
+} from "@/pdc/services/territory/contracts/findGeoBySiren.contract.ts";
+import { alias } from "@/pdc/services/territory/contracts/findGeoBySiren.schema.ts";
 import { blacklist } from "../../config/filterOutput.ts";
+import { GeoRepositoryProviderInterfaceResolver } from "../../interfaces/GeoRepositoryProviderInterface.ts";
 
 @handler({
   ...handlerConfig,
