+++ api/src/pdc/services/territory/actions/group/CreateTerritoryAction.ts
@@ -1,14 +1,10 @@
-import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { handler } from "@/ilos/common/index.ts";
+import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { hasPermissionMiddleware } from "@/pdc/providers/middleware/index.ts";
 
+import { handlerConfig, ParamsInterface, ResultInterface } from "@/pdc/services/territory/contracts/create.contract.ts";
+import { alias } from "@/pdc/services/territory/contracts/create.schema.ts";
 import { TerritoryRepositoryProviderInterfaceResolver } from "../../interfaces/TerritoryRepositoryProviderInterface.ts";
-import {
-  handlerConfig,
-  ParamsInterface,
-  ResultInterface,
-} from "@/shared/territory/create.contract.ts";
-import { alias } from "@/shared/territory/create.schema.ts";
 
 @handler({
   ...handlerConfig,
