+++ api/src/pdc/services/territory/actions/group/GetAuthorizedCodesAction.ts
@@ -1,14 +1,14 @@
-import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { handler } from "@/ilos/common/index.ts";
+import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { hasPermissionMiddleware } from "@/pdc/providers/middleware/index.ts";
 
-import { TerritoryRepositoryProviderInterfaceResolver } from "../../interfaces/TerritoryRepositoryProviderInterface.ts";
 import {
   handlerConfig,
   ParamsInterface,
   ResultInterface,
-} from "@/shared/territory/getAuthorizedCodes.contract.ts";
-import { alias } from "@/shared/territory/getAuthorizedCodes.schema.ts";
+} from "@/pdc/services/territory/contracts/getAuthorizedCodes.contract.ts";
+import { alias } from "@/pdc/services/territory/contracts/getAuthorizedCodes.schema.ts";
+import { TerritoryRepositoryProviderInterfaceResolver } from "../../interfaces/TerritoryRepositoryProviderInterface.ts";
 
 @handler({
   ...handlerConfig,
