+++ api/src/pdc/services/territory/actions/group/PatchContactsTerritoryAction.ts
@@ -1,17 +1,14 @@
 import { handler } from "@/ilos/common/index.ts";
 import { Action as AbstractAction } from "@/ilos/core/index.ts";
-import {
-  copyFromContextMiddleware,
-  hasPermissionByScopeMiddleware,
-} from "@/pdc/providers/middleware/index.ts";
+import { copyFromContextMiddleware, hasPermissionByScopeMiddleware } from "@/pdc/providers/middleware/index.ts";
 
-import { TerritoryRepositoryProviderInterfaceResolver } from "../../interfaces/TerritoryRepositoryProviderInterface.ts";
 import {
   handlerConfig,
   ParamsInterface,
   ResultInterface,
-} from "@/shared/territory/patchContacts.contract.ts";
-import { alias } from "@/shared/territory/patchContacts.schema.ts";
+} from "@/pdc/services/territory/contracts/patchContacts.contract.ts";
+import { alias } from "@/pdc/services/territory/contracts/patchContacts.schema.ts";
+import { TerritoryRepositoryProviderInterfaceResolver } from "../../interfaces/TerritoryRepositoryProviderInterface.ts";
 
 @handler({
   ...handlerConfig,
