+++ api/src/pdc/services/territory/contracts/common/interfaces/TerritoryInterface.ts
