+++ api/src/pdc/services/territory/interfaces/TerritoryRepositoryProviderInterface.ts
@@ -1,28 +1,28 @@
 import type {
   ParamsInterface as CreateParamsInterface,
   ResultInterface as CreateResultInterface,
-} from "@/shared/territory/create.contract.ts";
+} from "@/pdc/services/territory/contracts/create.contract.ts";
 
 import type {
   ParamsInterface as ListParamsInterface,
   ResultInterface as ListResultInterface,
-} from "@/shared/territory/list.contract.ts";
+} from "@/pdc/services/territory/contracts/list.contract.ts";
 
 import type {
   ParamsInterface as FindParamsInterface,
   ResultInterface as FindResultInterface,
-} from "@/shared/territory/find.contract.ts";
+} from "@/pdc/services/territory/contracts/find.contract.ts";
 
 import type {
   ParamsInterface as UpdateParamsInterface,
   ResultInterface as UpdateResultInterface,
-} from "@/shared/territory/update.contract.ts";
+} from "@/pdc/services/territory/contracts/update.contract.ts";
 
+import type { TerritorySelectorsInterface } from "@/pdc/services/territory/contracts/common/interfaces/TerritoryCodeInterface.ts";
 import type {
   ParamsInterface as PatchContactsParamsInterface,
   ResultInterface as PatchContactsResultInterface,
-} from "@/shared/territory/patchContacts.contract.ts";
-import type { TerritorySelectorsInterface } from "@/shared/territory/common/interfaces/TerritoryCodeInterface.ts";
+} from "@/pdc/services/territory/contracts/patchContacts.contract.ts";
 
 export type {
   CreateParamsInterface,
@@ -53,8 +53,7 @@ export interface TerritoryRepositoryProviderInterface {
   ): Promise<TerritorySelectorsInterface>;
 }
 
-export abstract class TerritoryRepositoryProviderInterfaceResolver
-  implements TerritoryRepositoryProviderInterface {
+export abstract class TerritoryRepositoryProviderInterfaceResolver implements TerritoryRepositoryProviderInterface {
   async list(params: ListParamsInterface): Promise<ListResultInterface> {
     throw new Error();
   }
